+++ CheckDigits.Net.Tests.Benchmarks/GlobalUsings.cs
@@ -1,6 +1,8 @@
 ﻿global using BenchmarkDotNet.Attributes;
 global using BenchmarkDotNet.Running;
 
+global using CheckDigits.Net.GeneralAlgorithms;
 global using CheckDigits.Net.Iso7064;
+global using CheckDigits.Net.ValueSpecificAlgorithms;
 global using CheckDigits.Net.Tests.Benchmarks;
