+++ CheckDigits.Net.Tests.Benchmarks/NhsAlgorithmValidateBenchmarks.cs
@@ -1,7 +1,5 @@
 ﻿// Ignore Spelling: Nhs
 
-using CheckDigits.Net.ValueSpecificAlgorithms;
-
 namespace CheckDigits.Net.Tests.Benchmarks;
 
 [MemoryDiagnoser]
