+++ CheckDigits.Net.Tests.Benchmarks/NpiAlgorithmValidateBenchmarks.cs
@@ -1,7 +1,5 @@
 ﻿// Ignore Spelling: Npi
 
-using CheckDigits.Net.ValueSpecificAlgorithms;
-
 namespace CheckDigits.Net.Tests.Benchmarks;
 
 [MemoryDiagnoser]
