+++ CheckDigits.Net.Tests.Unit/CheckDigits.Net.Tests.Unit.csproj
@@ -24,8 +24,4 @@
     <ProjectReference Include="..\CheckDigits.Net\CheckDigits.Net.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="GeneralAlgorithms\" />
-  </ItemGroup>
-
 </Project>
