+++ CheckDigits.Net.Tests.Unit/GlobalUsings.cs
@@ -2,6 +2,7 @@
 
 global using Xunit;
 
+global using CheckDigits.Net.GeneralAlgorithms;
 global using CheckDigits.Net.Iso7064;
 global using CheckDigits.Net.Utility;
 global using CheckDigits.Net.ValueSpecificAlgorithms;
