+++ CheckDigits.Net/GlobalUsings.cs
@@ -1,3 +1,4 @@
-﻿global using CheckDigits.Net.Iso7064;
+﻿global using CheckDigits.Net.GeneralAlgorithms;
+global using CheckDigits.Net.Iso7064;
 global using CheckDigits.Net.Utility;
 global using CheckDigits.Net.ValueSpecificAlgorithms;
